requirements.txt
# Core Backend
flask==3.0.0
pymongo==4.5.0

# Google Cloud Services
google-cloud-storage==2.10.0      # Cloud Storage
google-cloud-texttospeech==2.15.0 # Text-to-Speech
google-cloud-language==2.8.0      # NLP analysis
google-cloud-aiplatform==1.95.1   # AI Platform - Keep the latest one
google-cloud-secret-manager==2.16.1

# Data Processing
requests==2.31.0
Pillow==10.0.0
pandas==2.1.4                 # Optional
unidecode==1.3.6              # Optional

# Environment & Typing
python-dotenv==1.0.0
flask-cors==4.0.0
typing-extensions==4.8.0

# Retry Utility
tenacity==8.2.3
